requirements.txt
requests==2.22.0
gitpython==3.1.0
python-dateutil==2.8.1
lxml==4.5.0